pyproject.toml
[project]
name = "llm-agents"
version = "0.1.0"
description = "LLM-powered agents and arena for comparing multiple AI providers"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.76.0",
    "gradio>=6.5.1",
    "openai>=2.15.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "rich>=14.2.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.11.0",
]

[tool.pytest.ini_options]
asyncio_mode = "strict"
